% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_weighting_functions.R
\name{combine_weighting_functions}
\alias{combine_weighting_functions}
\title{Combine weight functions from two sampling frames}
\usage{
combine_weighting_functions(weights_function_1, weights_function_2)
}
\arguments{
\item{weights_function_1}{The weight function from the \emph{outer} sampling frame (the 'larger' scale; Records in one group of the \emph{outer} sampling frame can belong to different strata in the \emph{inner} sampling frame.)}

\item{weights_function_2}{The weight function from the \emph{inner} sampling frame (the 'smaller' scale; Records in the same group of the \emph{inner} sampling frame must also belong to the same group in the \emph{outer} sampling frame.)}
}
\description{
Combine weight functions from two sampling frames
}
\details{
The returned function combines two sets of weights so that
\itemize{
\item the sum of weights of each sampling frame's groups remain proportional to each other
\item the total sum of weights equals the number of rows in the input data frame
}
}
