% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weights.R
\name{stratify.count.sample}
\alias{stratify.count.sample}
\title{Get sample sizes for weight calculations.
Checks if all strata appear in sampling frame}
\usage{
stratify.count.sample(data.strata, sf.strata)
}
\description{
Get sample sizes for weight calculations.
Checks if all strata appear in sampling frame
}
