% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weights.R
\name{stratify.weights}
\alias{stratify.weights}
\title{estimates sampling probabilities from sample size and strata populations.}
\usage{
stratify.weights(pop_strata, sample_strata)
}
\description{
estimates sampling probabilities from sample size and strata populations.
}
