% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_fun_from_samplingframe.R
\name{weighting_fun_from_samplingframe}
\alias{weighting_fun_from_samplingframe}
\title{creates a weighting function from a sampling frame}
\usage{
weighting_fun_from_samplingframe(sampling.frame, data.stratum.column,
  sampling.frame.population.column = "population",
  sampling.frame.stratum.column = "stratum", data = NULL)
}
\arguments{
\item{data.stratum.column}{data column name that holds the record's strata names}

\item{sampling.frame.population.column}{sampling frame name of column holding population counts. defaults to "population"}

\item{sampling.frame.stratum.column}{sampling frame name of column holding stratum names. defaults to "stratum". Stratum names must match exactly values in:}

\item{data}{optional but recommended: you can provide an example data frame of data supposed to match the sampling frame to check if the provided variable names match and whether all strata in the data appear in the sampling frame.}

\item{sampling.frame.file}{data frame containing the sampling frame. should contain columns "stratum" and "population", otherwise column names must be specified.}
}
\value{
returns a new function that takes a data frame as input returns a vector of weights corresponding to each row in the data frame.
}
\description{
creates a weighting function from a sampling frame
}
\seealso{
\code{\link{load_data} (not implemented), \link{aggregate_count_weighted}}
}
